backend/requirements.txt
fastapi
uvicorn
pydantic
langchain-groq
langchain-community
sentence-transformers
chromadb
pymongo
python-dotenv
bcrypt
python-jose[cryptography]